`Refactor db connection and update package.json and index.js`
